requirements.txt: drop explicit version numbers
We are only using requirements.txt to bootstrap the development environment,
and that file is not really used in production. It is better to rely on
the latest and greatest so we actually follow pep8/pylint etc changes as
they happen.

If we get too many issues with that, we might pin some of the version
numbers in te future, but so far the experience has been that we simply
forget to bump versions and stick to very old versions instead.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>